add description in book mark

+++ app/assets/javascripts/book_marsk.js.coffee
@@ -1,11 +1,20 @@
 $(document).on 'ready page:load', ->
 
   $(document).on "blur", '.del input', (event) ->
-#    $(this).css 'background-color', 'yellow'
     value = $(this).val();
     input = $(event.target)
-    $.ajax
-      dataType: 'script'
-      type: 'DELETE'
-      data: {delete: value}
-      url: input.attr("data-load-event-url")
+
+    if value == 'del'
+      $.ajax
+        dataType: 'script'
+        type: 'DELETE'
+        data: {delete: value}
+        url: input.attr("data-load-event-url")
+
+    if value == 'edt'
+      edi_url = input.attr("data-load-event-url") + "/edit"
+      $.ajax
+        dataType: 'script'
+        type: 'GET'
+        data: {delete: value}
+        url: edi_url
